Load native libraries (resolves #7)
This is a breaking change. .NET assemblies must now be placed in libraries/managed, and native ones in libraries/native, as native libraries (which share the same file extension on Windows) are loaded by OS rather than CLR. Due of this, a call to NativeLibrary.Free is essential upon the service stopping!